+++ docs/changelog.rst
@@ -16,6 +16,8 @@ E7EPD Database Specification
     * Added Diode table
 * v0.3 (In Progress):
     * Added a ``storage`` key to all parts
+    * Added a ``for_project`` key to all parts
+    * Merged ``user_comments`` and ``part_comments`` to just one ``comments`` column
 
 Database Python DB Wrapper
 --------------------------------------------
@@ -30,6 +32,11 @@ Database Python DB Wrapper
     * Added ```wipe_database``` function
     * Added a way for the backend to store configurations about itself
     * Added a key to check the database specification the database is under and the ``E7EPD`` class
+* v0.3:
+    * Switched to sqlalchemy for handing SQL
+    * As the wrapper input includes a sqlalchemy engine, any sql type that sqlalchemy supports should be supported
+    * Updated spec for Database Rev 0.3
+    * Added the first migration from Database Rev 0.2 to 0.3 with ``alembic``
 
 CLI
 -----------
@@ -44,3 +51,8 @@ CLI
     * Added option to remove and append stock to a part
     * Moved around options so that there is an "initial screen" before choosing components
     * Added a check for the database revision on startup
+* v0.3 (In Progress)
+    * Updated for the new Wrapper 0.3 database argument
+    * Allowing option for a mySQL database
+    * Allowing option to add multiple databases
+    * Allowing option to select which database to connect to